# next-app/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aether Radar API",
    "version": "3.1.0",
    "description": "AI 生态雷达 v3.1 只读数据 API，用于工具、风险、场景、导出和报告查询。"
  },
  "servers": [
    {
      "url": "http://localhost:4173",
      "description": "local dev"
    }
  ],
  "paths": {
    "/api/status": {
      "get": {
        "summary": "数据状态",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/entities": {
      "get": {
        "summary": "工具/平台实体搜索",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "risk",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "curated",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "summary": "跨实体、词汇、场景搜索",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/scenarios": {
      "get": {
        "summary": "场景方案库",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/risks": {
      "get": {
        "summary": "风险维度与风险热点",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/export": {
      "get": {
        "summary": "导出数据",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "markdown",
                "xlsx"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/report": {
      "get": {
        "summary": "生成场景选型报告",
        "parameters": [
          {
            "name": "scenario",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "markdown"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/openapi": {
      "get": {
        "summary": "OpenAPI JSON",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    }
  }
}